READ ME , any golf fans out there?
Golf is a game in which very good golfers are asked to play in a PGA tournament or a certain Masters where their is winners, prestege and alot of money.  I love to get to watch and know my favorite players. Mine is Phil Michelson who is still playing but getting older but will go down in history as one of the most consistant greats of all time for his generation. Tiger Woods came into play around when Phil was playing and he holds records today that were absolutely unbelievable alot of people loved Tigerbut he fell on bad times and in my opinion I was glad I was always a Phil fan. All for now, add on if you want!!!

Ll